+++ src/main/java/open/commons/spring/web/config/ObjectMapperConfiguration.java
@@ -26,22 +26,15 @@
 
 package open.commons.spring.web.config;
 
-import java.util.Map;
-
+import javax.annotation.Nonnull;
 import javax.validation.constraints.NotNull;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
-import org.springframework.boot.autoconfigure.condition.ConditionalOnMissingBean;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Primary;
 import org.springframework.http.converter.json.Jackson2ObjectMapperBuilder;
-
-import open.commons.spring.web.jacksons.decoration.IObjectMapperDecorationConsolidator;
-import open.commons.spring.web.jacksons.decoration.IObjectMapperDecorator;
-import open.commons.spring.web.jacksons.decoration.ObjectMapperDecorationConsolidator;
-import open.commons.spring.web.jacksons.decoration.ObjectMapperDecorator;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 
@@ -58,53 +51,13 @@ public class ObjectMapperConfiguration {
 
     private final Logger logger = LoggerFactory.getLogger(ObjectMapperConfiguration.class);
 
-    /**
-     * <br>
-     * 
-     * <pre>
-     * [개정이력]
-     *      날짜    	| 작성자	|	내용
-     * ------------------------------------------
-     * 2025. 6. 17.		박준홍			최초 작성
-     * </pre>
-     *
-     *
-     * @since 2025. 6. 17.
-     * @version 0.8.0
-     * @author parkjunhong77@gmail.com
-     */
-    public ObjectMapperConfiguration() {
-    }
-
-    @Bean(ObjectMapperDecorator.BEAN_QUALIFIER)
-    @ConditionalOnMissingBean
-    IObjectMapperDecorator authorizedObjectMapperDecorator() {
-        IObjectMapperDecorator omd = new ObjectMapperDecorator();
-
-        logger.info("[object-mapper] object-mapper-decorator={}", omd);
-
-        return omd;
-    }
-
     @Bean(BEAN_QUALIFIER_DEFAULT_OBJECT_MAPPER)
     @Primary
-    ObjectMapper objectMapper(@NotNull IObjectMapperDecorationConsolidator objectMapperConsolidator) {
-        ObjectMapper mapper = Jackson2ObjectMapperBuilder.json().build();
-        objectMapperConsolidator.configureFeature(mapper);
+    ObjectMapper objectMapper(@NotNull @Nonnull Jackson2ObjectMapperBuilder objectMapperBuilder) {
+        ObjectMapper mapper = objectMapperBuilder.build();
 
         logger.info("[object-mapper] default-object-mapper={}", mapper);
 
         return mapper;
-    }
-
-    @Bean(ObjectMapperDecorationConsolidator.BEAN_QUALIFIER)
-    @Primary
-    IObjectMapperDecorationConsolidator objectMapperDecorationConsolidator(Map<String, IObjectMapperDecorator> decorators) {
-        IObjectMapperDecorationConsolidator omc = new ObjectMapperDecorationConsolidator();
-        omc.addObjectMapperDecorator(decorators.values());
-
-        logger.info("[object-mapper] object-mapper-decoration-consolidator={}", omc);
-
-        return omc;
     }
 }