+++ src/main/java/open/commons/spring/web/jacksons/decoration/ObjectMapperDecorator.java
@@ -29,7 +29,11 @@ package open.commons.spring.web.jacksons.decoration;
 import java.util.Collections;
 import java.util.Set;
 
+import org.springframework.context.annotation.Bean;
+import org.springframework.http.converter.json.Jackson2ObjectMapperBuilder;
+
 import com.fasterxml.jackson.databind.Module;
+import com.fasterxml.jackson.databind.ObjectMapper;
 import com.fasterxml.jackson.databind.jsontype.NamedType;
 
 /**
@@ -37,6 +41,9 @@ import com.fasterxml.jackson.databind.jsontype.NamedType;
  * @since 2025. 6. 17.
  * @version 0.8.0
  * @author parkjunhong77@gmail.com
+ * 
+ * @deprecated {@link Jackson2ObjectMapperBuilder}를 {@link Bean}으로 제공받아 {@link ObjectMapper}를 생성하는 방식으로 변경됨에 따라 필요성이
+ *             없어짐.
  */
 public class ObjectMapperDecorator implements IObjectMapperDecorator {